pyproject.toml
[project]
name = "gtd-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=1.8.0",
    "langchain-chroma>=0.2.5",
    "langchain-google-genai>=2.1.8",
    "langchain-huggingface>=0.3.1",
    "langchain-text-splitters>=0.3.9",
    "litellm>=1.74.8",
    "sentence-transformers>=5.0.0",
    "unstructured[md]>=0.18.11",
]
